save to hive

+++ LouvainModularityHive/src/main/scala/Driver.scala
@@ -38,7 +38,7 @@ object Driver {
     val louvainTbl = hc.table(config.hiveSchema + "." + config.hiveOutputTable)
     // register the table so it can be used in SQL
     louvainTbl.createOrReplaceTempView(config.hiveOutputTable)
-    val exists = hc.sql("select count(*) NrRecords from LouvainCommunity where MilanoDate = '" + config.dateInput + "' and alphaThreshold = " + config.alphaThreshold + " and edgeCostFactor =" + config.edgeCostFactor).first().getInt(0)
+    val exists = hc.sql("select count(*) NrRecords from LouvainCommunity where MilanoDate = '" + config.dateInput + "' and alphaThreshold = " + config.alphaThreshold + " and edgeCostFactor =" + config.edgeCostFactor).first().getLong(0)
     println("Value exists? " + exists)
     // val louvain = new Louvain()
     // louvain.run(sc, hc, config)
